Clima: Custom Delegate for WeatherManagerData

+++ Basic/Clima/Clima/Model/WeatherData.swift
@@ -20,5 +20,5 @@ struct Weather: Codable {
 }
 
 struct main: Codable {
-    var temp: Int
+    var temp: Double
 }
